% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncentral_t_normalization.R
\name{noncentral_t_normalization}
\alias{noncentral_t_normalization}
\title{Noncentral t-statistics normalization}
\usage{
noncentral_t_normalization(zv, v)
}
\arguments{
\item{zv}{A numeric vector of noncentral t-statistics.}

\item{v}{A numeric vector of the degrees of freedom.}
}
\value{
Normalized noncentral t-statistics.
}
\description{
An approximate normalization method for noncentral t-distributed z-values. The resulting normalized z-values will be approximately normally distributed with standard deviation 1.
}
