% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dStorey.R
\name{STS}
\alias{STS}
\title{Directional Storey, Taylor and Siegmund procedure.}
\usage{
STS(pv, alpha, B = 1000, lambdas = 0.5)
}
\arguments{
\item{pv}{A numeric of p-values.}

\item{alpha}{The directional false discovery rate target.}

\item{B}{Number of bootstrap samples.}

\item{lambdas}{A numeric of candidate \eqn{\lambda} values for the automatic directional Storey, Taylor and Siegmund procedure. If a single value is provided, then the standard directional Storey, Taylor and Siegmund procedure will run.}
}
\value{
The rejection indices.
}
\description{
An adaptive procedure that controls the directional false discovery rate by estimating the null proportion \eqn{\pi_0}.
}
