% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdirect.R
\name{zdirect}
\alias{zdirect}
\title{zdirect}
\usage{
zdirect(
  betahat,
  sebetahat,
  s_l = NULL,
  s_r = NULL,
  alpha = 0.05,
  mixcompdist = c("halfuniform", "halfnormal"),
  prior = "nullbiased",
  nullweight = 10,
  altweight = 0.9,
  nfits = 20,
  epsilon = 1e-10
)
}
\arguments{
\item{betahat}{A numeric of z-values \eqn{(z_1,\dots,z_m)}.}

\item{sebetahat}{A numeric of the standard deviations \eqn{(\sigma_1,\dots,\sigma_m)}.}

\item{s_l}{The left bound for \eqn{\mathcal{M}_1}. See below.}

\item{s_r}{The right bound for \eqn{\mathcal{M}_1} so that \eqn{ \mathcal{M}_1 = \{ i: u'_i \leq \texttt{s\_l} \text{ or } u'_i \geq \texttt{s\_r} \} }.}

\item{alpha}{The directional false discovery rate target.}

\item{mixcompdist}{The approximation model \eqn{\hat{G}(\cdot)} for \eqn{G(\cdot)}. If \eqn{\texttt{"halfuniform"}} is chosen, then \eqn{\hat{G}(\cdot)} will be a mixture of uniform distributions. If \eqn{\texttt{"halfnormal"}} is chosen, then \eqn{\hat{G}(\cdot)} will be a mixture of mixture of truncated normal distributions.}

\item{prior}{A string, or numeric vector indicating Dirichlet prior on mixture proportions: \eqn{\texttt{"nullbiased"}}, \eqn{\texttt{c(nullweight,altweight,...,altweight)}}, puts more weight on first component; \eqn{\texttt{"uniform"}} is \eqn{\texttt{c(1,1...,1)}}; \eqn{\texttt{"unit"}} is \eqn{\texttt{c(1/K,...,1/K)}}.}

\item{nullweight}{A scalar, the weight put on the prior for nulls under "nullbiased" specification, see \eqn{\texttt{prior}}.}

\item{altweight}{A scalar, the weight put on the prior for non-nulls under "nullbiased" specification, see \eqn{\texttt{prior}}.}

\item{nfits}{A scalar that ensures that \eqn{\hat{g}_t(\cdot)} is only re-estimated every \eqn{\lceil m / \texttt{nfits} \rceil} steps where \eqn{m} is the length of \eqn{\texttt{betahat}}.}

\item{epsilon}{A small scalar which widens the width of the non-masked data set at each iteration.}
}
\value{
The rejection indices.
}
\description{
A procedure that controls the directional false discovery rate via "data masking" for z-values of the form \eqn{z_i \sim \text{N}(\theta,\sigma_i)}. If the z-values available are noncentral t-distributed, \eqn{\texttt{noncentral\_t\_normalization}} should be used to approximately normalize them first before inputting them into \eqn{\texttt{zdirect}}.
}
